Trait raw_sync_2::locks::LockInit

pub trait LockInit {
    // Required methods
    fn size_of(addr: Option<*mut u8>) -> usize;
    unsafe fn new(
        mem: *mut u8,
        data: *mut u8
    ) -> Result<(Box<dyn LockImpl>, usize), Box<dyn Error>>;
    unsafe fn from_existing(
        mem: *mut u8,
        data: *mut u8
    ) -> Result<(Box<dyn LockImpl>, usize), Box<dyn Error>>;
}

Required Methods

fn size_of(addr: Option<*mut u8>) -> usize

Size required for the lock’s internal representation

unsafe fn new( mem: *mut u8, data: *mut u8 ) -> Result<(Box<dyn LockImpl>, usize), Box<dyn Error>>

Initializes a new instance of the lock in the provided buffer and returns the number of used bytes

Safety

This function is unsafe because it cannot guarantee that the provided memory is valid.

unsafe fn from_existing( mem: *mut u8, data: *mut u8 ) -> Result<(Box<dyn LockImpl>, usize), Box<dyn Error>>

Re-uses a lock from an already initialized location and returns the number of used bytes

Safety

This function is unsafe because it cannot guarantee that the provided memory is valid.

Object Safety

This trait is not object safe.